+++ TrackForce/src/main/java/com/revature/utils/PersistentStorage.java
@@ -44,7 +44,7 @@ public class PersistentStorage {
 	}
 	
 	//added method to actually get values from the cache
-	public AssociateInfo getAssociate(BigDecimal id) {
+	public AssociateInfo getAssociate(Integer id) {
 		return associates.get(id);
 	}
 
@@ -71,7 +71,7 @@ public class PersistentStorage {
 	}
 
 	//added method to actually get values from the cache
-	public ClientInfo getClient(BigDecimal id) {
+	public ClientInfo getClient(Integer id) {
 		return clients.get(id);
 	}
